api_yamdb/users/models.py
import jwt

from datetime import datetime, timedelta

from django.contrib.auth.models import (AbstractUser, BaseUserManager,
                                        PermissionsMixin)
from django.conf import settings
from django.db import models

CHOICES = (
    ('user', 'Обычный пользователь'),
    ('moderator', 'Модератор'),
    ('admin', 'Администратор'),
)


class UserManager(BaseUserManager):
    """
    Django требует, чтобы кастомные пользователи определяли свой собственный
    класс Manager. Унаследовавшись от BaseUserManager, мы получаем много того
    же самого кода, который Django использовал для создания User (для демонстрации).
    """

    def create_user(self, username, email, password=None, confirmation_code='',
                    role='user', bio=''):
        """ Создает и возвращает пользователя с имэйлом, паролем и именем. """
        if username is None:
            raise TypeError('Users must have a username.')

        if email is None:
            raise TypeError('Users must have an email address.')

        user = self.model(username=username, email=self.normalize_email(email),
                          confirmation_code=confirmation_code, role=role, bio=bio)
        user.set_password(password)
        user.save()

        return user

    def create_superuser(self, username, email, password, role='user', bio=''):
        """ Создает и возввращет пользователя с привилегиями суперадмина. """
        if password is None:
            raise TypeError('Superusers must have a password.')

        user = self.create_user(username, email, password, role=role, bio=bio)
        user.is_superuser = True
        user.is_staff = True
        user.save()

        return user


class User(AbstractUser):
    ADMIN = 'admin'
    MODERATOR = 'moderator'
    USER = 'user'
    CHOISES = (
        (ADMIN, 'Administrator'),
        (MODERATOR, 'Moderator'),
        (USER, 'User'),
    )
    username = models.CharField(db_index=True, max_length=255, unique=True)
    email = models.EmailField(db_index=True, unique=True)
    bio = models.TextField(
        'Информация о пользователе',
        max_length=1000,
        blank=True,
    )
    role = models.CharField(
        'Роль пользователя',
        max_length=30,
        choices=CHOICES,
        default=USER,
    )
    #is_moderator = models.BooleanField(
    #    'Модератор',
    #    default=False,
    #)
    USERNAME_FIELD = 'email'
    REQUIRED_FIELDS = ['username']
    #confirmation_code = models.CharField(
    #    'Код подтверждения',
    #    max_length=10
    #)
    #objects = UserManager()
    @property
    def is_moderator(self):
        return self.role == self.MODERATOR
    
    @property
    def is_admin(self):
        return self.role == self.ADMIN

    #def __str__(self):
    #    """ Строковое представление модели (отображается в консоли) """
    #    return self.username

    @property
    def token(self):
        """
        Позволяет получить токен пользователя путем вызова user.token, вместо
        user._generate_jwt_token(). Декоратор @property выше делает это
        возможным. token называется "динамическим свойством".
        """
        return self._generate_jwt_token()

    def _generate_jwt_token(self):
        """
        Генерирует веб-токен JSON, в котором хранится идентификатор этого
        пользователя, срок действия токена составляет 1 день от создания
        """
        dt = datetime.now() + timedelta(days=1)

        token = jwt.encode({
            'user_id': self.pk,
            'exp': dt.utcfromtimestamp(dt.timestamp())
        }, settings.SECRET_KEY, algorithm='HS256')

        return token.decode('utf-8')
